libnm: drop unrechable g_warn_if_fail() assertion from NMSettingVpn's update_secret_dict()
For one, this code was unreachable, because we checked these conditions
already before.

That aside, g_warn_*() is for all intended purposes an assertion.
The caller probably gets the GVariant from an untrusted source (e.g. via
D-Bus). Asserting here is not helpful.
It's up to the caller to validate the argument. Or, in case the caller
doesn't care, update_secret_dict() should just do the sensible thing.
But be silent about it!