libnm: refactor property setters of data and secrets in NMSettingVpn
Don't use _nm_utils_copy_strdict().

On a minor note, that function will always allocate a GHashTable, even if
it only contains "" as only key. Later we would throw that out again,
so it was unnecessary.

Worse, using _nm_utils_copy_strdict() does not use nm_free_secrets as
destroy function. While it's in general difficult to clear all places
in memory where we copy the secrets around, we can easily avoid that.

Also skip over %NULL keys and values. It probably would be a bug passing
such arguments to the property. Better ignore them and not crash.