cli: use _value_strsplit() in _set_fcn_optionlist()
VALUE_STRSPLIT_MODE_OBJLIST already uses the right delimiter (","). It also
strips the whitespace, just like we anyway want.

Use _value_strsplit().

This is not particularly useful, but the goal is to unify our ways how
we split the string. Use _value_strsplit() with a suitable option
instead of re-implementing it.