cli: support backslash escaping for cli options like vpn.data, vpn.secrets, bond.options, ethernet.s390-options
This is obviously a change in behavior, was we now honor backslash
escape sequences. With this, all string values can be expressed.
Previously, you could for example not set vpn.secrets to have a ','
and you could not set vpn.data to

  nmcli connection modify "$PROFILE" +vpn.data 'ipsec-ike = aes256-sha2_256-modp2048,aes256-sha2_256-modp1536'

Use a relatively simple backslash escaping scheme. The main goal of
the scheme is that it doesn't change behavior for the majority of case.
It only changes behavior if:

  - your value contains a backslash

  - if the backslash proceeds one of the few characters that support
    escaping now. That is, whitespace, ',', and '\'.
    For the key part (before the first '='), the '=' also requires
    escaping, while for the value part the escape sequecen '\=' is
    taken literally.

The downside is that in most cases backslashes are treated as regular
characters, unless the preceed a special character that supports
escaping. Especially confusing is that '\=' is interpreted differently
for the key and value part.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/issues/ ## 390